ARM: OMAP: smsc911x: add support for MAC address
Current gpmc-smsc911x has no support for passing MAC address to the
driver. In case the SMSC911x chip does not have a connected storage with
MAC address, this results in a randomly generated MAC address.
This patch allows board code to pass the MAC address via the platform
data.