INPUT: ads7846 - fix wrong transfer init for ads7845
For ads7845 transfer packets are wrong initialized for
activated settle_delay_usec. Buffers of spi_transfers have to point
to the ads7845 buffers also for the transfers used for implementing
the settling delay. Otherwise the filter will see invalid / uninitialized
data and the input gets the first raw samples.

This patch fixes the wrong buffer init.

Signed-off-by: Markus Niebel <Markus.Niebel@tqs.de>