ARM: OMAP3: cm-t35: add USB hub reset GPIO pin mux
Pin mux was not setup for the sb-t35/cb-t3 USB hub reset pin.
It works for sb-t35 as there is a pull up on that pin.
cb-t3 does not have a pull up on the same pin and therefore the USB hub
does not work without the patch.
Make USB hub work on both base boards.

Signed-off-by: Ilya Ledvich <ilya@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>